(#17) Parse incoming messages skipping the payload
Without this commit, the parser loads the AST of the entire message in memory
and then traverses it to return its elements. This commit makes the parser able to
read the message as a stream but skipping the payload. The parser returns a lazy
structure which could be used by clients to lazily read arguments.